+++ lib/generators/pg_sql_triggers/templates/create_pg_sql_triggers_tables.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreatePgTriggersTables < ActiveRecord::Migration[6.0]
+class CreatePgSqlTriggersTables < ActiveRecord::Migration[6.0]
   def change
     # Registry table - source of truth for all triggers
     create_table :pg_sql_triggers_registry do |t|